+++ src/main/java/io/inhyuck/webservice/config/WebSecurityConfig.java
@@ -5,7 +5,6 @@
 
 package io.inhyuck.webservice.config;
 
-import io.inhyuck.webservice.domain.member.MemberAuthenticationProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
