+++ src/main/java/io/inhyuck/webservice/dto/DeveloperResumeResponseDto.java
@@ -1,7 +1,6 @@
 package io.inhyuck.webservice.dto;
 
-import io.inhyuck.webservice.domain.resume.Card;
-import io.inhyuck.webservice.domain.resume.ResumeDevelop;
+import io.inhyuck.webservice.entity.resume.ResumeDevelop;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
