+++ src/main/java/io/inhyuck/webservice/entity/member/Member.java
@@ -3,7 +3,7 @@
  * Author: inhyuck | https://github.com/inhyuck
  */
 
-package io.inhyuck.webservice.domain.member;
+package io.inhyuck.webservice.entity.member;
 
 import lombok.Builder;
 import lombok.Data;
