+++ src/main/java/io/inhyuck/webservice/entity/resume/UrlList.java
@@ -1,4 +1,4 @@
-package io.inhyuck.webservice.domain.resume;
+package io.inhyuck.webservice.entity.resume;
 
 import lombok.Data;
