+++ src/main/java/io/inhyuck/webservice/model/member/LoginForm.java
@@ -3,7 +3,7 @@
  * Author: inhyuck | https://github.com/inhyuck
  */
 
-package io.inhyuck.webservice.domain.member;
+package io.inhyuck.webservice.model.member;
 
 import lombok.Data;
