+++ src/main/java/io/inhyuck/webservice/service/MemberService.java
@@ -6,7 +6,7 @@
 package io.inhyuck.webservice.service;
 
 import io.inhyuck.webservice.dao.MemberDAO;
-import io.inhyuck.webservice.domain.member.Member;
+import io.inhyuck.webservice.entity.member.Member;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
