+++ src/main/java/io/inhyuck/webservice/service/ResumeService.java
@@ -3,9 +3,9 @@ package io.inhyuck.webservice.service;
 import io.inhyuck.webservice.component.CardComponent;
 import io.inhyuck.webservice.config.MySheetProperties;
 import io.inhyuck.webservice.dao.ResumeDAO;
-import io.inhyuck.webservice.domain.resume.Card;
-import io.inhyuck.webservice.domain.resume.ResumeDesign;
-import io.inhyuck.webservice.domain.resume.ResumeDevelop;
+import io.inhyuck.webservice.dto.Card;
+import io.inhyuck.webservice.entity.resume.ResumeDesign;
+import io.inhyuck.webservice.entity.resume.ResumeDevelop;
 import io.inhyuck.webservice.dto.DesignerResumeResponseDto;
 import io.inhyuck.webservice.dto.DeveloperResumeResponseDto;
 import io.inhyuck.webservice.dto.ResumeListResponseDto;
